Content_Generation_Engine/requirements.txt
groq
pytrends
python-dotenv
markdown